+++ lib/openapi_first/test/app.rb
@@ -10,16 +10,24 @@ module OpenapiFirst
     # A wrapper of the original app
     # with silent request/response validation to track requests/responses.
     class App < SimpleDelegator
-      def initialize(app, api:)
+      def initialize(app, api:, validate_request_before_handling:)
         super(app)
         @app = app
         @definition = Test[api]
+        @validate_request_before_handling = validate_request_before_handling
       end
 
       def call(env)
         request = Rack::Request.new(env)
-        env[Test::REQUEST] = @definition.validate_request(request, raise_error: false)
+        if @validate_request_before_handling
+          env[Test::REQUEST] = @definition.validate_request(request, raise_error: false)
+        end
+
         response = @app.call(env)
+        unless @validate_request_before_handling
+          env[Test::REQUEST] = @definition.validate_request(request, raise_error: false)
+        end
+
         status, headers, body = response
         env[Test::RESPONSE] =
           @definition.validate_response(request, Rack::Response[status, headers, body], raise_error: false)